manifests: set command, volume and volume mounts acording to platform
For the OpenShift platform we will use notify file created by OCI hook
to monitor when RTE should update the topology resource.

For the kubernetes we will use /var/lib/kubelet directory to monitor
when RTE should update the topology resource and also to provide the
kubelet config to the RTE daemon set.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>